Moved selection implementations to module
