+++ holoviews/element/chart.py
@@ -5,7 +5,8 @@ from ..streams import BoundsXY
 from ..core import util
 from ..core import Dimension, Dataset, Element2D
 from ..core.data import GridInterface
-from .geom import Selection2DExpr, Rectangles, Points, VectorField # noqa: backward compatible import
+from .geom import Rectangles, Points, VectorField # noqa: backward compatible import
+from .selection import Selection1DExpr, Selection2DExpr
 
 
 class Chart(Dataset, Element2D):
@@ -45,92 +46,6 @@ class Chart(Dataset, Element2D):
 
     def __getitem__(self, index):
         return super(Chart, self).__getitem__(index)
-
-
-class Selection1DExpr(Selection2DExpr):
-    """
-    Mixin class for Cartesian 1D Chart elements to add basic support for
-    SelectionExpr streams.
-    """
-
-    _selection_dims = 1
-
-    _inverted_expr = False
-
-    def _get_selection_expr_for_stream_value(self, **kwargs):
-        from ..util.transform import dim
-        from ..core import NdOverlay
-        from .annotation import HSpan, VSpan
-
-        invert_axes = self.opts.get('plot').kwargs.get('invert_axes', False)
-        region_el = HSpan if invert_axes or self._inverted_expr else VSpan
-        if kwargs.get('bounds', None) is None:
-            region = None if 'index_cols' in kwargs else NdOverlay({0: region_el()})
-            return None, None, region
-
-        x0, y0, x1, y1 = kwargs['bounds']
-
-        # Handle invert_xaxis/invert_yaxis
-        if y0 > y1:
-            y0, y1 = y1, y0
-        if x0 > x1:
-            x0, x1 = x1, x0
-
-        if len(self.dimensions()) == 1:
-            xdim = self.dimensions()[0]
-            ydim = None
-        else:
-            xdim, ydim = self.dimensions()[:2]
-
-        if invert_axes:
-            x0, x1, y0, y1 = y0, y1, x0, x1
-            xdim, ydim = ydim, xdim
-        if self._inverted_expr:
-            if ydim is not None: xdim = ydim
-            x0, x1 = y0, y1
-
-        bbox = {xdim.name: (x0, x1)}
-        index_cols = kwargs.get('index_cols')
-        if index_cols:
-            index_cols = [self.get_dimension(c) for c in index_cols]
-            sel = self.dataset.select(**bbox)
-            vals = dim(index_cols[0], util.unique_zip, *index_cols[1:]).apply(sel)
-            selection_expr = dim(index_cols[0], util.lzip, *index_cols[1:]).isin(vals)
-            region_element = None
-        else:
-            selection_expr = ((dim(xdim) >= x0) & (dim(xdim) <= x1))
-            region_element = NdOverlay({0: region_el(x0, x1)})
-        return selection_expr, bbox, region_element
-
-    @staticmethod
-    def _merge_regions(region1, region2, operation):
-        from ..core import NdOverlay
-        if region1 is None or operation == "overwrite":
-            return region2
-        data = [d.data for d in region1] + [d.data for d in region2]
-        prev = len(data)
-        new = None
-        while prev != new:
-            prev = len(data)
-            contiguous = []
-            for l, u in data:
-                if not util.isfinite(l) or not util.isfinite(u):
-                    continue
-                overlap = False
-                for i, (pl, pu) in enumerate(contiguous):
-                    if l >= pl and l <= pu:
-                        pu = max(u, pu)
-                        overlap = True
-                    elif u <= pu and u >= pl:
-                        pl = min(l, pl)
-                        overlap = True
-                    if overlap:
-                        contiguous[i] = (pl, pu)
-                if not overlap:
-                    contiguous.append((l, u))
-            new = len(contiguous)
-            data = contiguous
-        return NdOverlay([(i, region1.last.clone(l, u)) for i, (l, u) in enumerate(data)])
 
 
 class Scatter(Selection1DExpr, Chart):
