+++ holoviews/element/stats.py
@@ -5,8 +5,7 @@ from ..core.dimension import Dimension, process_dimensions
 from ..core.data import Dataset
 from ..core.element import Element, Element2D
 from ..core.util import get_param_values, unique_iterator, OrderedDict
-from .chart import Selection1DExpr
-from .geom import Selection2DExpr
+from .selection import Selection1DExpr, Selection2DExpr
 
 
 class StatisticsElement(Dataset, Element2D):
